[cm6] Basic Rocq language mode for CM6
This is a basic mode for Rocq + CM6.

We re-use the stream parser we had for CM5, with a few small
differences as we don't need to use the parser now to determine the
end of the sentence thanks to Flèche.

We should also define a mode for `.mv` files, this should be kind of
easy using CM, we can re-use the markdown mode and define Rocq
Markdown as a sub-language.

This a first step, a lot more nice things can be done with CodeMirror.